Refactor
Falta implementar el \PUT, la resta funciona
