+++ routes/auth.js
@@ -1,0 +1,8 @@
+var express = require('express')
+var router = express.Router()
+var jwt = require('jsonwebtoken')
+var authCtrl = require('../controllers/auth')
+
+exports.login = function (req, res, next) {
+    authCtrl.login(req, res, next)
+}
